+++ src/jobs/handlers/notification_handler.rs
@@ -180,6 +180,7 @@ mod tests {
                 paused: false,
                 policies: Some(RelayerNetworkPolicyResponse::Evm(
                     RelayerEvmPolicy {
+                        include_revert_data: None,
                         gas_price_cap: None,
                         whitelist_receivers: None,
                         eip1559_pricing: None,