+++ src/repositories/relayer/relayer_in_memory.rs
@@ -282,6 +282,7 @@ mod tests {
             paused: false,
             network_type: NetworkType::Evm,
             policies: RelayerNetworkPolicy::Evm(RelayerEvmPolicy {
+                include_revert_data: None,
                 gas_price_cap: None,
                 whitelist_receivers: None,
                 eip1559_pricing: Some(false),
@@ -464,6 +465,7 @@ mod tests {
 
         // Create a new policy to update
         let new_policy = RelayerNetworkPolicy::Evm(RelayerEvmPolicy {
+            include_revert_data: None,
             gas_price_cap: Some(50000000000),
             whitelist_receivers: Some(vec!["0x1234".to_string()]),
             eip1559_pricing: Some(true),